refactor(gui): move the Load/LazyDeviceData infrastructure to state/load.rs
The lazy load-state types have no AppState dependency; state.rs keeps
the state itself.
